Ejemplo con enemigo básico y animación de combate. s

+++ Assets/Scripts/Health.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class Health : MonoBehaviour
 {
-    public int health = 10;
+    public float health = 10;
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         health -= damage;
         if(health <= 0)
